requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
big-O==0.9
bleach==3.1.0
cachetools==4.0.0
certifi==2019.6.16
chardet==3.0.4
click==7.1.1
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
dialogflow==0.8.0
entrypoints==0.3
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.12.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
multidict==4.5.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
opencv-python==4.2.0.32
pandas==1.0.1
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.3
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.5
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyrsistent==0.15.6
PySimpleGUI==4.18.0
pytesseract==0.3.3
python-dateutil==2.8.0
python-launch-library==0.6
pytz==2019.3
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.3
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
twilio==6.37.0
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yarl==1.3.0
zipp==0.6.0
